backend/requirements.txt
fastapi
uvicorn
httpx
openai
python-dotenv
pycountry
google-generativeai
slowapi
together